monarch_hyperactor

Module alloc

Structs

PyAlloc

A python class that wraps a Rust Alloc trait object. It represents what is shown on the python side. Internals are not exposed. It ensures that the Alloc is only used once (i.e. moved) in rust.

PyAllocConstraints

PyAllocSpec

PyLocalAllocator

PyProcessAllocator

PyRemoteAllocator

PyRemoteProcessAllocInitializer

A [hyperactor_mesh::alloc::RemoteProcessAllocInitializer] wrapper to enable subclassing from Python.

PySimAllocator

Functions

register_python_bindings