monarch_hyperactor::code_sync

Module rsync

Structs

Change

Represents a single file change from rsync

RsyncActor

RsyncDaemon

RsyncMessage

RsyncResult

Represents the result of an rsync operation with details about what was transferred

Enums

ChangeAction

ChangeMessage

ChangeType

The type of change that occurred to a file

FileType

The type of file that changed

Functions

do_rsync

rsync_mesh