monarch_hyperactor

Module proc

Structs

InstanceWrapper

Wrapper around an instance of an actor that provides utilities to implement a python actor. This helps by allowing users to specialize the actor to the message type they want to handle.

PyActorId

PyProc

Wrapper around a proc that provides utilities to implement a python actor.

PySerialized

Wrapper around a [Any] that allows returning it to python and passed to python based detached actors to send to other actors.

Functions

register_python_bindings